+++ src/services/index.ts
@@ -1,0 +1,1 @@
+export { default as GameService } from './game/game.service';
